\name{LOCIDSO}
\docType{data}
\alias{LOCIDSO}
\title{Animal capture details file}
\description{
This file contains individual animal capture histories and locations
}
\usage{data(LOCIDSO)}
\format{csv}
\source{SPACECAP reference manual}
\references{
Royle, J. A., K. U. Karanth, A. M. Gopalaswamy and N. S. Kumar. 2009. Bayesian inference in camera trapping studies
  for a class of spatial capture-recapture models. \emph{Ecology} \bold{90(11)}, 3233-3244.
}
\keyword{datasets}