\name{LOCSO}
\docType{data}
\alias{LOCSO}
\title{Trap deployment details file}
\description{
This file provides information on the dates when each camera trap location was
active and operational during the survey
}
\usage{data(LOCSO)}
\format{csv}
\source{SPACECAP reference manual}
\references{
Royle, J. A., K. U. Karanth, A. M. Gopalaswamy and N. S. Kumar. 2009. Bayesian inference in camera trapping studies
  for a class of spatial capture-recapture models. \emph{Ecology} \bold{90(11)}, 3233-3244.
}
\keyword{datasets}