\name{tigerhab}
\docType{data}
\alias{tigerhab}
\title{Potential home range centers}
\description{
This file describes potential animal home range center details in terms of their spatial
location and habitat suitability indicator for these home range centers
}
\usage{data(tigerhab)}
\format{csv}
\source{SPACECAP reference manual}
\references{
Royle, J. A., K. U. Karanth, A. M. Gopalaswamy and N. S. Kumar. 2009. Bayesian inference in camera trapping studies
  for a class of spatial capture-recapture models. \emph{Ecology} \bold{90(11)}, 3233-3244.
}
\keyword{datasets}
