\name{SPACECAP-package}
\alias{SPACECAP-package}
\docType{package}
\title{A Program to Estimate Animal Abundance and Density using Bayesian Spatially-Explicit Capture-Recapture Models}
\description{
    \pkg{SPACECAP} is a user-friendly software package for estimating animal 
    densities using closed model capture-recapture sampling based 
    on photographic captures using Bayesian spatially-explicit capture- 
    recapture models. This approach offers advantage such 
    as: substantially dealing with problems posed by individual 
    heterogeneity in capture probabilities in conventional 
    capture-recapture analyses. It also offers non-asymptotic 
    inferences which are more appropriate for small samples of 
    capture data typical of photo-capture studies.
}
\details{
\tabular{ll}{
Package: \tab SPACECAP\cr
Type: \tab Package\cr
Version: \tab 1.0.6\cr
Date: \tab 2012-06-04\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Arjun M. Gopalaswamy, Jeffrey A. Royle, James E. Hines, Pallavi Singh, Devcharan Jathanna, N. Samba Kumar and K. Ullas Karanth 
}
\references{
Royle, J. A., K. U. Karanth, A. M. Gopalaswamy and N. S. Kumar. 2009. Bayesian inference in camera trapping studies 
  for a class of spatial capture-recapture models. \emph{Ecology} \bold{90(11)}, 3233-3244. 
}
